Move processContentType to common.go and use for creation as well
